Rename BeatmapCoverAvatar to AlbumImage
