+++ src/components/beatmap/cover/AlbumImage.vue
